feat: Migrate menu and lobby UI to feature controllers and i18n
