+++ src/i18n/locales/uk.json
@@ -1,6 +1,7 @@
 {
   "common": {
     "back": "Назад",
+    "cancel": "Скасувати",
     "close": "Закрити",
     "language": "Мова",
     "loading": "Завантаження...",
@@ -9,10 +10,112 @@
   },
   "menu": {
     "actions": {
+      "continue": "Продовжити",
       "exit": "Вийти",
-      "play": "Грати"
+      "play": "Грати",
+      "quit": "Вийти",
+      "saveGame": "Зберегти гру",
+      "saving": "Збереження..."
     },
+    "confirmBackToMain": "Повернутися до головного меню?",
     "loadingLobbies": "Відкриття лобі..."
+  },
+  "lobby": {
+    "actions": {
+      "createLobby": "Створити лобі",
+      "hostFirstLobby": "Створити перше лобі",
+      "hostLobby": "Створити лобі",
+      "inGame": "У грі",
+      "joinLobby": "Приєднатися",
+      "leaveLobby": "Покинути лобі",
+      "mainMenu": "Головне меню",
+      "refresh": "Оновити",
+      "refreshSaves": "Оновити збереження",
+      "startGame": "Почати гру"
+    },
+    "description": {
+      "browser": "Переглядайте відкриті світи, збереження або створюйте нову сесію без втручання в ігровий рендерер.",
+      "warTable": "Зберіть групу, перевірте джерело сесії та почніть забіг, коли хост буде готовий."
+    },
+    "dialog": {
+      "lobbyName": "Назва лобі",
+      "lobbyNamePlaceholder": "Введіть назву зали",
+      "saveSlot": "Слот збереження",
+      "sessionSource": "Джерело сесії",
+      "title": "Створити лобі"
+    },
+    "empty": {
+      "description": "Створіть першу кімнату або оновіть список після того, як інший гравець створить її. У цій версії відкриті лише лобі в очікуванні.",
+      "title": "Немає доступних лобі"
+    },
+    "fields": {
+      "joinRule": "Правило входу",
+      "origin": "Походження",
+      "players": "Гравці"
+    },
+    "hostLabel": "Хост: {{label}}",
+    "hostModes": {
+      "loadSave": "Завантажити збереження",
+      "newGame": "Нова гра"
+    },
+    "joinRule": {
+      "locked": "Закрито",
+      "open": "Відкрито"
+    },
+    "loadedFrom": "Завантажено з {{name}}",
+    "loadingWorld": {
+      "description": "Підключення ігрового сокета та рендерера світу без виходу з активної ігрової сесії.",
+      "eyebrow": "Фаза сесії",
+      "title": "Завантаження світу"
+    },
+    "memberCount": "{{count}} учасник{{suffix}}",
+    "messages": {
+      "contactingRegistry": "Зв'язок з реєстром лобі"
+    },
+    "origin": {
+      "loadedSave": "Завантажене збереження",
+      "newGame": "Нова гра",
+      "newShort": "Нова",
+      "saveShort": "Збереження"
+    },
+    "playEyebrow": "Гра",
+    "presence": {
+      "inWorld": "У світі",
+      "waiting": "Очікує"
+    },
+    "roles": {
+      "host": "Хост",
+      "member": "Учасник",
+      "you": "Ви"
+    },
+    "saves": {
+      "defaultName": "{{lobbyName}}: збереження",
+      "empty": "Серверних збережень ще немає.",
+      "slotLabel": "Слот збереження",
+      "sourceLobby": "Джерело лобі: {{name}}",
+      "updated": "Оновлено {{date}}",
+      "version": "v{{version}}"
+    },
+    "sections": {
+      "session": "Сесія",
+      "sessionSource": "Джерело сесії"
+    },
+    "source": {
+      "activeSaveBound": "Активний слот збереження прив'язаний до збережень у сесії",
+      "description": "Лобі в очікуванні доступні для входу у v1. Після старту сесії ігрові сокети підключаються до цього світу, а нові входи блокуються.",
+      "freshWorld": "Новий світ",
+      "freshWorldDescription": "Це лобі створено з нового авторитетного екземпляра світу.",
+      "loadedSave": "Завантажене збереження"
+    },
+    "status": {
+      "closed": "Закрито",
+      "in_game": "У грі",
+      "waiting": "Очікування"
+    },
+    "title": {
+      "browser": "Перегляд лобі",
+      "warTable": "Військова рада"
+    }
   },
   "settings": {
     "languageNames": {
